requirements.txt
# Core dependencies for protein evolution using MSA-Transformer
torch>=1.10.0
numpy>=1.21.0
scipy>=1.7.0

# ESM (Evolutionary Scale Modeling) - MSA-Transformer and ESM2
fair-esm>=2.0.0

# Bioinformatics tools
biopython>=1.79
pysam>=0.19.0

# Utilities
tqdm>=4.62.0
heapq>=0.1.0

# Optional: Experiment tracking and visualization
# wandb>=0.12.0
# matplotlib>=3.5.0
# seaborn>=0.11.0
